services/snapshot/requirements.txt
txt
opencv-python-headless
Flask
firebase-admin
google-cloud-storage
gunicorn
numpy